Subtasks can now be marked as completed
Felt a little hacky. I had to modify several different pieces of state in order for the UI to reflect that the subtask had been marked completed, and for the checkbox to be checked. I expected that upon updated the meta board data, the changes would be reflected throughout the entire UI. Maybe I need to approach the architecture a little differently in future.