Bug fix
Changing subtask completion status on other boards caused a null reference error. This was due to the findNestedObject function not containing functionality to account for the object being of an array type. I set up some extra conditions to account for this situation, and added an array-specific iteration/recursion method